+++ core/src/main/java/com/main/GameManager.java
@@ -8,104 +8,31 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import javax.swing.*;
 import java.awt.*;
 
-public class GameManager implements Screen {
-    private SpriteBatch batch;
+public class GameManager {
+    Menu menu;
+    String currScreen;
 
-    private Texture background;
-    private Texture title;
-
-    private Texture btnPlay;
-    private Texture btnSetting;
-    private Texture btnScore;
-    private Texture btnQuit;
+    public GameManager(String screen){
+        menu = new Menu();
 
 
 
-    public GameManager(){
+        currScreen = screen;
+    }
 
-        //load ảnh từ file
-        background = new Texture("Stuffs/background.png");
+    public void draw() {
 
-        //load ảnh chữ ( tên game )
-        title = new Texture("Stuffs/Title.png");
+        switch (currScreen) {
+            case "menu":
+//                vẽ menu
+                menu.draw();
+                break;
 
-        //Load ảnh các nút
-        btnPlay = new Texture("Stuffs/Buttons/Play.png");
-        btnSetting = new Texture("Stuffs/Buttons/Setting.png");
-        btnScore = new Texture("Stuffs/Buttons/Score.png");
-        btnQuit = new Texture("Stuffs/Buttons/Exit.png");
-
-        // cấu hình
-        batch = new SpriteBatch();
+            case "game":
+//                vẽ game
+                break;
+        }
     }
 
 
-    @Override
-    public void show() {
-
-    }
-
-    @Override
-    public void render(float v) {
-        batch.begin();
-        // vẽ nền background
-        batch.draw(background, 0, 0, Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
-
-        // vẽ chữ hiện trên nền background
-        batch.draw(title, 150, Gdx.graphics.getHeight() - title.getHeight() - 20);
-
-        // vị trí ban đầu để vẽ các nút
-        int btnX = 10; // Vị trí x của nút, dịch chuyển sang bên trái hơn
-        int currentY = Gdx.graphics.getHeight() - 250; // Bắt đầu từ trên cao, dưới tiêu đề
-        int spacing = 20; // Khoảng cách giữa các nút
-        int buttonWidth = 250; // Chiều rộng nút
-        int buttonHeight = 80; // Chiều cao nút
-
-        // Vẽ nút PLAY
-        batch.draw(btnPlay, btnX, currentY, buttonWidth, buttonHeight);
-        currentY -= buttonHeight + spacing;
-
-        // Vẽ nút SETTING
-        batch.draw(btnSetting, btnX, currentY, buttonWidth, buttonHeight);
-        currentY -= buttonHeight + spacing;
-
-        // Vẽ nút SCORE
-        batch.draw(btnScore, btnX, currentY, buttonWidth, buttonHeight);
-        currentY -= buttonHeight + spacing;
-
-        // Vẽ nút EXIT
-        batch.draw(btnQuit, btnX, currentY, buttonWidth, buttonHeight);
-        batch.end();
-    }
-
-    @Override
-    public void resize(int i, int i1) {
-
-    }
-
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void hide() {
-
-    }
-
-    @Override
-    public void dispose() {
-        background.dispose();
-        title.dispose();
-        btnPlay.dispose();
-        btnSetting.dispose();
-        btnScore.dispose();
-        btnQuit.dispose();
-        batch.dispose();
-    }
 }
