+++ lwjgl3/src/main/java/com/main/lwjgl3/Lwjgl3Launcher.java
@@ -30,7 +30,7 @@ public class Lwjgl3Launcher {
 
 
         configuration.setWindowedMode(1344, 672);
-
+        configuration.setResizable(false);
         //// You can change these files; they are in lwjgl3/src/main/resources/ .
         //// They can also be loaded from the root of assets/ .
         configuration.setWindowIcon("libgdx128.png", "libgdx64.png", "libgdx32.png", "libgdx16.png");
